show grayscale,blur,normal image from API on btn click fixed

+++ app/src/main/java/com/arditakrasniqi/mymobilelife/data/model/Image.kt
@@ -8,10 +8,10 @@ import java.net.URL
 data class Image(
     val id: String,
     val author: String,
-    val width: Long,
-    val height: Long,
+    val width: Int,
+    val height: Int,
     val url: URL,
     val download_url: URL
-): Parcelable
+) : Parcelable
